ci: add Dependabot configuration file
This commit introduces a 'dependabot.yml' configuration file for automated dependency management. The configuration ensures daily updates of both GitHub Actions and PNPM dependencies. Additionally, it specifies that pull requests for updates should be assigned to 'nekofar', labelled as 'dependencies', limited to a maximum of 20 for GitHub Actions and 20 for PNPM, reviewed by 'nekofar', and targeted to the 'develop' branch.