ci: add stale issue handler configuration file
This commit introduces a .github/stale.yml configuration file to handle issues that have become stale. The config sets stale to be at 60 days of inactivity and issues will be closed after being stale for 7 days. It also sets 'pinned' and 'security' labeled issues as exempted from being marked as stale. It labels stale issues as 'wontfix' and automatically leaves a comment when an issue is marked as stale.