chore(deps-dev): add `husky`, `lint-staged`, and `prettier` plugins to `pnpm-lock.yaml`
In this commit, a set of devDependencies - husky, lint-staged, and multiple prettier plugins - are added to the pnpm-lock.yaml file. These tools are commonly used for enforcing code style rules and automated pre-commit tasks, enhancing the overall code quality and consistency across the project.

Additionally, the URLs for downloading several packages' tarballs are removed from the resolution field, leaving only the integrity hash. It is unnecessary to specify the tarball link when the package manager can directly retrieve it from the registry using the package name and version. Removing the links eases maintenance and updates.