chore: add `turbo.json` for new Turbo setup
This commit introduces a new 'turbo.json' file associated with the Turbo build system. The addition of the JSON file provides a predefined structure for build, lint and development process for enhancing management. The two distinct output paths, '.next' and '!next/cache' are included in the 'build' pipeline for better organization of built assets. The default caching is disabled in the 'dev' pipeline for increased development efficiency.