chore(deps-dev): update TypeScript version from 5.2.2 to 5.3.2 in package dependencies
Version of TypeScript used in the project was updated from 5.2.2 to 5.3.2 in 'pnpm-lock.yaml' file. This commit introduces several changes in dependencies that use TypeScript, updating the version in their respective lines. The TypeScript update is necessary to take advantage of the latest features, improvements, and bug fixes provided in the new version.