chore: add dependencies to `turbo.json` and new command to `package.json`
Changes are made in turbo.json to add dependencies to the build and pages:build pipelines. This was necessary to ensure that the lint runs before the build starts for both pipelines. In addition, a new command was added to package.json, "turbo:pages:build" to allow building turbo pages as needed. The addition of these dependencies and commands is designed to enhance the flow of commands and ensure correct building procedures.