refactor: enforce production mode in mdx serialize function
Changes made to serialization code in the index.tsx, vote/index.tsx, and [slug].tsx files to enforce production mode. This is done by setting `development: false` in the mdxOptions. These changes were made to prevent possible bugs in development phase that might not appear in production due to the discrepancies between the processing of mdx files in different environments.