refactor: allow dynamic imports for specified node modules
Dynamic imports have been allowed for lodash and walletconnect node modules in multiple API endpoints. This change allows for the usage of these packages within serverless functions on edge environment. As a result, these node modules will now be included in the bundle which will enhance the performance and versatility of the application. It will also eliminate the need for static imports, which can lead to faster loading times and improved user experience.