refactor: adapt `mdxOptions` based on `NODE_ENV`
Changed the 'development' option in the mdxOptions object to toggle dynamically based on the NODE_ENV. This will let the application behave differently depending on the environment it is running in, enhancing the flexibility and adaptability when testing versus production. This will effortlessly improve the quality control in development process as well.