refactor: update API handler import and add edge runtime
In `src/pages/api/[[...route]].ts`, the import path of the 'handle' function was updated from `@hono/node-server/vercel` to `hono/vercel`. This change was necessary to align with recent updates in the package. In addition, 'runtime' was specified as 'edge' to improve performance by allowing the code to run closer to the user.