refactor: add `tsconfig.json` file for function settings
Introduce a new tsconfig.json file under the 'functions' directory. This file contains the configuration for the TypeScript compiler. ESNext is set as target and module, enabling the most recent JavaScript features. Additionally, the "strict" flag is activated for enforcing stricter type-checking, promoting a more robust codebase. The "@cloudflare/workers-types" and "hono/jsx" are included to leverage the libraries for the Cloudflare worker environment and JSX syntax respectively. Furthermore, path aliases are created for cleaner and more readable import statements.