refactor: update environment variables handling in wallet configurations
This commit improves the handling of environment variables in the wallet configs. Instead of directly accessing process.env in multiple occurrences for `NEXT_PUBLIC_TOKEN_NETWORK` value and `NEXT_PUBLIC_ALCHEMY_KEY` value, we have created a constant called `network` and `alchemyKey` at the beginning of the file. This change ensures that process.env is not undefined and directly increases reliability and readability. The same mechanism is also applied to the TOKEN_CONTRACT and ETHER_ACTOR_BASEURL constants in the addresses.ts and urls.ts files.